+++ src/main/java/com/illtamer/perpetua/sdk/event/notice/FriendNoticeEvent.java
@@ -25,7 +25,7 @@ public abstract class FriendNoticeEvent extends NoticeEvent {
      * 向该消息发送者发送消息
      * @return 消息 ID
      * */
-    public Long sendMessage(String message) {
+    public Integer sendMessage(String message) {
         return OpenAPIHandling.sendMessage(message, userId);
     }
 
@@ -33,7 +33,7 @@ public abstract class FriendNoticeEvent extends NoticeEvent {
      * 向该消息发送者发送消息
      * @return 消息 ID
      * */
-    public Long sendMessage(Message message) {
+    public Integer sendMessage(Message message) {
         return OpenAPIHandling.sendMessage(message, userId);
     }
